requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
click==8.3.2
colorama==0.4.6
fastapi==0.135.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
lxml==6.0.4
openai==2.32.0
pydantic==2.13.0
pydantic_core==2.46.0
python-dotenv==1.2.2
soupsieve==2.8.3
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
WeasyPrint==65.1
motor==3.3.2
pymongo==4.6.1
